+++ app/uk/gov/hmrc/eventhub/respository/SubscriberEventRepository.scala
@@ -16,49 +16,12 @@
 
 package uk.gov.hmrc.eventhub.respository
 
-import cats.data.OptionT
-import cats.implicits._
-import org.mongodb.scala.model.Filters.equal
-import play.api.Logging
-import uk.gov.hmrc.eventhub.model.{ Event, SubscriberWorkItem }
-import uk.gov.hmrc.mongo.workitem.WorkItem
+import uk.gov.hmrc.eventhub.model.Event
 
-import scala.concurrent.{ ExecutionContext, Future }
+import scala.concurrent.Future
 
 trait SubscriberEventRepository {
   def next(): Future[Option[Event]]
   def failed(event: Event): Future[Option[Boolean]]
   def sent(event: Event): Future[Option[Boolean]]
-}
-
-class WorkItemSubscriberEventRepository(
-  subscriberQueueRepository: SubscriberQueueRepository
-)(implicit executionContext: ExecutionContext)
-    extends SubscriberEventRepository with Logging {
-  override def next(): Future[Option[Event]] =
-    subscriberQueueRepository.getEvent
-      .map(_.map(_.item.event))
-
-  override def failed(event: Event): Future[Option[Boolean]] =
-    (for {
-      workItem <- OptionT(findAsWortItem(event))
-      result   <- OptionT(subscriberQueueRepository.failed(workItem).map(_.some))
-    } yield {
-      logger.info(s"marking $event as failed: $result")
-      result
-    }).value
-
-  override def sent(event: Event): Future[Option[Boolean]] =
-    (for {
-      workItem <- OptionT(findAsWortItem(event))
-      result   <- OptionT(subscriberQueueRepository.completeAndDelete(workItem.id).map(_.some))
-    } yield {
-      logger.info(s"marking $event as sent: $result")
-      result
-    }).value
-
-  private def findAsWortItem(event: Event): Future[Option[WorkItem[SubscriberWorkItem]]] =
-    subscriberQueueRepository.collection
-      .find(equal("item.event.eventId", event.eventId.toString))
-      .headOption()
 }
